+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.25.0] - 2024-09-23
+
 ### Added
 
 - Added several cluster-wide WaitConditions for Deployments, DaemonSets, StatefulSets, Jobs and Pods being in an expected state.
@@ -597,7 +599,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added initial framework layout
 - Added Kubernetes client extended from controller-runtime client
 
-[Unreleased]: https://github.com/giantswarm/clustertest/compare/v1.24.1...HEAD
+[Unreleased]: https://github.com/giantswarm/clustertest/compare/v1.25.0...HEAD
+[1.25.0]: https://github.com/giantswarm/clustertest/compare/v1.24.1...v1.25.0
 [1.24.1]: https://github.com/giantswarm/clustertest/compare/v1.24.0...v1.24.1
 [1.24.0]: https://github.com/giantswarm/clustertest/compare/v1.23.0...v1.24.0
 [1.23.0]: https://github.com/giantswarm/clustertest/compare/v1.22.0...v1.23.0